+++ src/sputil/core/Base64.cpp
@@ -112,9 +112,9 @@ static inline bool is_base64(uint8_t c)
     return (isalnum(c) || (c == '+') || (c == '/'));
 }
 
-static int internal_decode(Buffer& dest, std::string const& encoded_string)
+static size_t internal_decode(Buffer& dest, std::string const& encoded_string)
 {
-    int in_len = encoded_string.size();
+    size_t in_len = encoded_string.size();
     int i = 0;
     int in_ = 0;
     uint8_t char_array_4[4], char_array_3[3];
@@ -125,7 +125,7 @@ static int internal_decode(Buffer& dest, std::string const& encoded_string)
     char_array_4[i++] = encoded_string[in_]; in_++;
     if (i ==4) {
       for (i = 0; i <4; i++)
-        char_array_4[i] = base64_chars.find(char_array_4[i]);
+        char_array_4[i] = (uint8_t) base64_chars.find(char_array_4[i]);
 
       char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
       char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -143,7 +143,7 @@ static int internal_decode(Buffer& dest, std::string const& encoded_string)
       char_array_4[j] = 0;
 
     for (j = 0; j <4; j++)
-      char_array_4[j] = base64_chars.find(char_array_4[j]);
+      char_array_4[j] = (uint8_t) base64_chars.find(char_array_4[j]);
 
     char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
     char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
